app/requirements.txt
gql[all]>=3.0.0
aiohttp>=3.8
websockets>=11.0
SQLAlchemy>=2.0
asyncpg>=0.27
python-dotenv
streamlit
pandas
psycopg2-binary
pymongo>=4.0.0
fastapi
uvicorn[standard]